requirements.txt
# Backend API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
sqlmodel>=0.0.22
python-dotenv>=1.0.0
aiosqlite>=0.19.0
httpx>=0.23.0


# Client Desktop
requests>=2.31.0
pystray>=0.19.0
customtkinter>=5.2.2
Pillow>=10.0.0
